cambio url

+++ src/app/services/project.service.ts
@@ -19,9 +19,9 @@ export class ProjectsService {
 
 
   getProjectsByClient(client: number): Observable<any>{
-    return this.httpClient.get(`${environment.apiUrl}/projects/client/${this.authorization.getId()}`)
+    //return this.httpClient.get(`${environment.apiUrl}/projects/client/${this.authorization.getId()}`)
 
-    //return this.httpClient.get(`${environment.apiUrl}/client/${this.authorization.getId()}/projects/`) // cambiamos la URL y ponemos el cliente delante
+    return this.httpClient.get(`${environment.apiUrl}/clients/${this.authorization.getId()}/projects/`) // cambiamos la URL y ponemos el cliente delante
   }
   
 // //creamos la función del observador
@@ -31,17 +31,17 @@ export class ProjectsService {
     console.log('cliente', this.client);
 
     // console.log(this.httpClient.get(`${environment.apiUrl}/projects`).toPromise());
-    return this.httpClient.get(`${environment.apiUrl}/projects/`).toPromise();
+    return this.httpClient.get(`${environment.apiUrl}/clients/${this.authorization.getId()}/projects/`).toPromise();
   }
 
   async deleteProject(id: number) {
 
-    return this.httpClient.delete(`${environment.apiUrl}/projects/${id}`).toPromise();
+    return this.httpClient.delete(`${environment.apiUrl}/clients/${this.authorization.getId()}/projects/${id}`).toPromise();
     //this.projectsArray = this.projectsArray.filter(b => b.id !== id);
   }
 
   async editProject(id: number, body) {
-    return this.httpClient.put(`${environment.apiUrl}/projects/${id}`, body).toPromise();
+    return this.httpClient.put(`${environment.apiUrl}/clients/${this.authorization.getId()}/projects/${id}`, body).toPromise();
 
     //  this.projectsArray = this.projectsArray.map(b => {
     //    if(b.id === id){
@@ -52,11 +52,11 @@ export class ProjectsService {
   }
 
   async addProject(project) {
-    return this.httpClient.post(`${environment.apiUrl}/projects/`, project).toPromise();
+    return this.httpClient.post(`${environment.apiUrl}/clients/${this.authorization.getId()}/projects/`, project).toPromise();
   }
 
   async getProjectById(id) {
-    return this.httpClient.get(`${environment.apiUrl}/projects/${id}`).toPromise();
+    return this.httpClient.get(`${environment.apiUrl}/clients/${this.authorization.getId()}/projects/${id}`).toPromise();
     //return this.projectsArray.find(elem => elem.id === id);
   }
 }
